Flask support for user login and account creation
Users can now "log in" or sign up with a new account. Log in checks for username in system, TODO: add password check as well. Sign up requires a unqiue username, and if such a username is entered, the user's data in the form is entered into the database as a new player entry. Additionally, attempts to access the homepage while not logged in will redirect to the login screen, and attempts to access the login or sign up screens while logged in will redirect to the homepage