'My Campaigns' page designed & implemented
The 'My Campaigns' page now displays a (scrollable, if too many) list of boxes featuring a 'preview' of that campaign - the campaign's name and the name of the associated adventuring party - which when clicked pass the campaign ID value to the 'Campaign Details' page. On the DB end, a view linking campaigns to their characters and players, and a related stored procedure to access a selection of those values based on the logged in player have been added to facilitate this page's functionality. IN PROGRESS/TODO: password requirements from last commit, Campaign Details functionality&design