Initial foundation for Search functionality and minor miscellaneous fixes
Main changes: implementation of rough foundation for Search query generation. Search page now presents a dropdown list of searchable entities to user, which on selection generate in the same page a list of all fields for that entity along with an appropriate entry area for that field, ie. textfield for varchar/int, dropdown for enum, dropdown for foreign key value possibilities, etc. TODO: actual generation of query from these fields and display of results; CSS formatting of Search page
Minor changes: Implemented password requirements for users and logins. Minor column additions to a selection of entities to support the "display name" criteria for non-associative entities.