Implemented Search Results functionality
When search conditions are submitted, the appropriate select statement is formed, with values returned and printed based on the "preview" columns set for the given entity.  Fixed a bug where setting a value for a foreign key in the entity yielded unsearchable values. Added a stub for entity details, which the options on the Search Results page will ultimately route to. TODO: all html and formatting for search.html and search_results.html