Create Entity Functionality
Players can now create Campaigns, Characters, or MonsterParties. On selecting an option, the player is presented a page with entries for all built-in fields of the entity, as well as dynamically add-able drop downs for all associated entities linked to that base entity. When the player hits "Create" on the page, all entered values along with their liked attribute name are submitted to the database in a series of ordered insert_record calls, inserting the data into the database.