Private & Public Campaign Views support and Add Associative Entity Implemented for Existing Entities
Created views for public and private campaigns, and implemented code to differentiate which view is returned based on whether the caller is the game's dm or not. Additionally, reworked existing code for inserting associative records when creating a new entity to allow for adding single-level and multi-level static and dynamic associative entities to existing base entity records.